artemis

+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/internal/RPCDriver.kt
@@ -212,7 +212,7 @@ data class RPCDriverDSL(
                     "${RPCApi.RPC_CLIENT_QUEUE_NAME_PREFIX}.#" to AddressSettings().apply {
                         maxSizeBytes = maxBufferedBytesPerClient
                         addressFullMessagePolicy = AddressFullMessagePolicy.PAGE
-                        pageSizeBytes = maxSizeBytes / 10
+                        pageSizeBytes = (maxSizeBytes / 10).toInt()
                     }
             )
         }
